# monotickets-api/docs/contracts/OPENAPI_monotickets.yaml
openapi: 3.0.0
paths:
...
  /events/{eventId}/guests:
    post:
      operationId: GuestsController_create
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestDto"
      responses:
        "201":
          description: Guest created successfully
        "404":
          description: Event not found
      security: &a5
        - bearer: []
      summary: Create a new guest for an event
      tags: &a6
        - guests
    get:
      operationId: GuestsController_findAll
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Guests retrieved successfully
      security: *a5
      summary: Get all guests for an event
      tags: *a6
  /events/{eventId}/guests/bulk:
    post:
      operationId: GuestsController_bulkCreate
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUploadDto"
      responses:
        "201":
          description: Guests created successfully
          content:
            application/json:
              schema:
                example:
                  message: Successfully created 50 guests
                  count: 50
                  guests: []
        "400":
          description: Invalid CSV data
        "404":
          description: Event not found
      security: *a5
      summary: Bulk create guests (CSV upload)
      tags: *a6
  /events/{eventId}/guests/stats:
    get:
      operationId: GuestsController_getStats
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Guest stats retrieved successfully
          content:
            application/json:
              schema:
                example:
                  totalGuests: 150
                  confirmedGuests: 120
                  pendingGuests: 25
                  declinedGuests: 5
                  totalGuestCount: 300
      security: *a5
      summary: Get guest statistics for an event
      tags: *a6
  /guests/{id}:
    get:
      operationId: GuestsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Guest retrieved successfully
        "404":
          description: Guest not found
      security: *a5
      summary: Get guest by ID
      tags: *a6
    patch:
      operationId: GuestsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuestDto"
      responses:
        "200":
          description: Guest updated successfully
        "404":
          description: Guest not found
      security: *a5
      summary: Update guest
      tags: *a6
    delete:
      operationId: GuestsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Guest deleted successfully
        "404":
          description: Guest not found
      security: *a5
      summary: Delete guest
      tags: *a6
...
  /guest/invitation/{token}:
    get:
      operationId: GuestController_getInvitation
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Invitation retrieved successfully
          content:
            application/json:
              schema:
                example:
                  invitation:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    qrToken: EV-550E8400-E29B41D4-ABC123-DEF456
                    status: SENT
                  guest:
                    id: 550e8400-e29b-41d4-a716-446655440001
                    fullName: Juan Pérez
                    guestCount: 2
                    rsvpStatus: PENDING
                  event:
                    id: 550e8400-e29b-41d4-a716-446655440002
                    name: Boda de Juan y María
                    date: 2025-12-25T18:00:00.000Z
                    location: Jardín Las Rosas, CDMX
                    description: Celebración de nuestra boda
        "404":
          description: Invitation not found
      summary: Get invitation details (public)
      tags: &a11
        - guest
  /guest/rsvp:
    post:
      operationId: GuestController_confirmRsvp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpDto"
      responses:
        "200":
          description: RSVP confirmed successfully
          content:
            application/json:
              schema:
                example:
                  message: RSVP confirmed successfully
                  guest:
                    id: 550e8400-e29b-41d4-a716-446655440001
                    fullName: Juan Pérez
                    rsvpStatus: CONFIRMED
        "404":
          description: Invitation not found
      summary: Confirm RSVP (public)
      tags: *a11
  /guest/qr/{token}:
    get:
      operationId: GuestController_getQR
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: QR code retrieved successfully
          content:
            application/json:
              schema:
                example:
                  qrToken: EV-550E8400-E29B41D4-ABC123-DEF456
                  guestName: Juan Pérez
                  message: QR code retrieved successfully
        "404":
          description: QR code not available
      summary: Get QR code (public, only for confirmed guests)
      tags: *a11
  /guest/calendar/{token}:
    get:
      operationId: GuestController_getCalendar
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Calendar file generated successfully
          headers:
            Content-Type:
              description: text/calendar
            Content-Disposition:
              description: attachment; filename="event.ics"
        "404":
          description: Invitation not found
      summary: Download calendar file .ics (public)
      tags: *a11
...
  /exports/guests/{eventId}:
    get:
      operationId: ExportsController_exportGuests
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: CSV generated successfully
      security: &a16
        - bearer: []
      summary: Export guests list (CSV)
      tags: &a17
        - exports
  /exports/rsvp/{eventId}:
    get:
      operationId: ExportsController_exportRsvp
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: CSV generated successfully
      security: *a16
      summary: Export RSVP responses (CSV)
      tags: *a17
  /exports/attendance/{eventId}:
    get:
      operationId: ExportsController_exportAttendance
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: CSV generated successfully
      security: *a16
      summary: Export attendance scans (CSV)
      tags: *a17
...
info:
  title: Monotickets Platinum API
  description: API oficial para Monotickets Platinum (invited events, RSVPs,
    delivery y scanner).
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateGuestDto:
      type: object
      properties:
        fullName:
          type: string
          description: Guest full name
          example: Juan Pérez García
        phone:
          type: string
          description: Guest phone number
          example: +52 55 1234 5678
        email:
          type: string
          description: Guest email
          example: juan.perez@example.com
        guestCount:
          type: number
          description: Number of guests (including main guest)
          example: 2
          minimum: 1
          maximum: 10
        rsvpStatus:
          type: string
          description: RSVP status
          enum: &a21
            - PENDING
            - CONFIRMED
            - DECLINED
          example: PENDING
      required:
        - fullName
        - guestCount
    BulkUploadDto:
      type: object
      properties:
        guests:
          description: Array of guests to create
          type: array
          items:
            $ref: "#/components/schemas/CreateGuestDto"
      required:
        - guests
    UpdateGuestDto:
      type: object
      properties:
        fullName:
          type: string
          description: Guest full name
          example: Juan Pérez García
        phone:
          type: string
          description: Guest phone number
          example: +52 55 1234 5678
        email:
          type: string
          description: Guest email
          example: juan.perez@example.com
        guestCount:
          type: number
          description: Number of guests (including main guest)
          example: 2
          minimum: 1
          maximum: 10
        rsvpStatus:
          type: string
          description: RSVP status
          enum: *a21
          example: PENDING
...
    RsvpDto:
      type: object
      properties:
        qrToken:
          type: string
          description: QR token from invitation
          example: EV-550E8400-E29B41D4-ABC123-DEF456
        status:
          type: string
          description: RSVP status
          enum:
            - PENDING
            - CONFIRMED
            - DECLINED
          example: CONFIRMED
      required:
        - qrToken
        - status
...